cleaned up unused files
