--- client/src/utils/API.js
@@ -1,30 +1,0 @@
-// route to get logged in user's info (needs the token)
-export const getMe = (token) => {
-    return fetch('/api/users/me', {
-      headers: {
-        'Content-Type': 'application/json',
-        authorization: `Bearer ${token}`,
-      },
-    });
-  };
-  
-  export const createUser = (userData) => {
-    return fetch('/api/users', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(userData),
-    });
-  };
-  
-  export const loginUser = (userData) => {
-    return fetch('/api/users/login', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify(userData),
-    });
-  };
-
